+++ DoffenSSHTunnel/source/CustomButtonDialog.cpp
@@ -2,6 +2,7 @@
 #include "CustomButtonDialog.h"
 #include "CustomActionsEditWidget.h"
 #include "PasswordDb.h"
+#include "TruncOrViewportClipToolTipDelegate.h"
 
 CustomButtonDialog::CustomButtonDialog(QWidget *parent, ATSkeletonWindow *pMainWindow)
 : QDialog(parent)
@@ -16,6 +17,7 @@ CustomButtonDialog::CustomButtonDialog(QWidget *parent, ATSkeletonWindow *pMainW
 
 	editCommand->setIgnoreReturnKey(true);
 
+    tableVarHelp->setItemDelegate(new TruncOrViewportClipToolTipDelegate(tableVarHelp));
 	tableVarHelp->setContextMenuPolicy(Qt::CustomContextMenu);
 	m_pTableVarContextMenu = new QMenu(tableVarHelp);
 	m_pTableVarActionCopy = new QAction("&Copy (Ctrl+C)",m_pTableVarContextMenu);
@@ -367,6 +369,7 @@ void CustomButtonDialog::addHeaderToVarTable(const QString & strHeader)
 {
 	QTableWidgetItem *item = new QTableWidgetItem();
 	item->setText(strHeader);
+    item->setToolTip(strHeader); // Tooltip is only shown then text is truncated or clipped by viewport (thanks to TruncOrViewportClipToolTipDelegate)
 	item->setFont(QFont(QString::fromUtf8("Tahoma"),8, QFont::Bold));
 	int row = tableVarHelp->rowCount();
 	tableVarHelp->insertRow(row);
@@ -378,6 +381,7 @@ void CustomButtonDialog::addEmptyToVarTable()
 {
 	QTableWidgetItem *item = new QTableWidgetItem();
 	item->setText("empty");
+    item->setToolTip("empty"); // Tooltip is only shown then text is truncated or clipped by viewport (thanks to TruncOrViewportClipToolTipDelegate)
 	item->setFont(QFont(QString::fromUtf8("Tahoma"),8, QFont::Cursive));
 	int row = tableVarHelp->rowCount();
 	tableVarHelp->insertRow(row);
@@ -394,6 +398,7 @@ void CustomButtonDialog::addDataToVarTable(const QList<VariableStruct> & varList
 
 		QTableWidgetItem *itemName = new QTableWidgetItem();
 		itemName->setText(var.strName);
+        itemName->setToolTip(itemName->text()); // Tooltip is only shown then text is truncated or clipped by viewport (thanks to TruncOrViewportClipToolTipDelegate)
 		tableVarHelp->setItem(row,0,itemName);
 
 		QTableWidgetItem *itemValue = new QTableWidgetItem();
@@ -402,10 +407,12 @@ void CustomButtonDialog::addDataToVarTable(const QList<VariableStruct> & varList
         } else {
             itemValue->setText(var.strValue);
         }
+        itemValue->setToolTip(itemValue->text()); // Tooltip is only shown then text is truncated or clipped by viewport (thanks to TruncOrViewportClipToolTipDelegate)
 		tableVarHelp->setItem(row,1,itemValue);
 
 		QTableWidgetItem *itemDescr = new QTableWidgetItem();
 		itemDescr->setText(var.strDescription);
+        itemDescr->setToolTip(itemDescr->text()); // Tooltip is only shown then text is truncated or clipped by viewport (thanks to TruncOrViewportClipToolTipDelegate)
 		tableVarHelp->setItem(row,2,itemDescr);
 	}
 }
@@ -420,10 +427,12 @@ int CustomButtonDialog::addDataToVarTable(Tunnel_c* pt, const PortForwardStruct&
 
 	QTableWidgetItem *itemName = new QTableWidgetItem();
 	itemName->setText(pfs.strName);
+    itemName->setToolTip(itemName->text()); // Tooltip is only shown then text is truncated or clipped by viewport (thanks to TruncOrViewportClipToolTipDelegate)
 	tableVarHelp->setItem(row,0,itemName);
 
 	QTableWidgetItem *itemValue = new QTableWidgetItem();
 	itemValue->setText(QString("%1").arg(pfs.nLocalPort));
+    itemValue->setToolTip(itemValue->text()); // Tooltip is only shown then text is truncated or clipped by viewport (thanks to TruncOrViewportClipToolTipDelegate)
 	tableVarHelp->setItem(row,1,itemValue);
 
 	QString descr = "";
@@ -433,6 +442,7 @@ int CustomButtonDialog::addDataToVarTable(Tunnel_c* pt, const PortForwardStruct&
 	descr += pfs.strDescription;
 	QTableWidgetItem *itemDescr = new QTableWidgetItem();
 	itemDescr->setText(descr);
+    itemDescr->setToolTip(itemDescr->text()); // Tooltip is only shown then text is truncated or clipped by viewport (thanks to TruncOrViewportClipToolTipDelegate)
 	tableVarHelp->setItem(row,2,itemDescr);
 
 	return 1;