+++ DoffenSSHTunnel/source/TableView.cpp
@@ -1,9 +1,10 @@
 #include "TableView.h"
+#include "TruncOrViewportClipToolTipDelegate.h"
 
 TableView::TableView(QWidget *parent)
     : QTableView(parent)
 {
-
+    this->setItemDelegate(new TruncOrViewportClipToolTipDelegate(this));
 }